fix: quote windows autostart paths
